Update linux-setup.sh to work with ubuntu 20.04
Summary:
These changes only work 20.04 so we'll want to make this more
robust to work on more systems.  I'm only putting this up as
a proof-of-concept diff.  I ended up having to ditch the VM
since it was too slow to run webapp and run the tests.  This is
because it was splitting system resources with MacOS.  I'd like
to use Linux in the future for dev work, but I need to get an
actual Linux system first.

Issue: none

Test Plan:
- make
- see that it installs all the packages we need

Reviewers: csilvers, eleith, davidbraley

Reviewed By: csilvers, eleith, davidbraley

Subscribers: davidbraley

Differential Revision: https://phabricator.khanacademy.org/D63977