Uninstall go-backports on machines that don't support it.
Before, we were adding the go-backports ppa and just not using it if
we didn't need it (ubuntu 20+ doesn't need it).

BUt it turns out that ppa is actively harmful on ubuntu 20; after
adding it `apt-get upgrade` gives this error:
> E: The repository 'http://ppa.launchpad.net/longsleep/golang-backports/ubuntu focal Release' does not have a Release file.

So now we actually remove the ppa if `apt-get upgrade` fails after
adding it.

Auditors: kevinb

Test plan:
`make install` on ubuntu 20 (focal)