Update virtualenv
Summary:
In D63977 I update the linux config to support ubunut 20.04.  As part of
this change I updated the linux setup to install the current version of
virtualenv and update the create_and_activate_virtualenv() shared function
appropriately to use the --always-copy option.  I unfortunately did not
test this in MacOS and this option is not supported with the older version
of virutalenv that setup.sh was installing.  This diff fixes that by ensuring
that the version of virtualenv used between the two is the same.

I tried upgrading to 20.0.28 which is the most recent version of virtualenv,
but ran into the following issue when trying to create a virtualenv:
    ImportError: cannot import name enquote_executable

I installed 20.0.23 which is what I had installed previously and that works
like a charm.

Issue: none

Test Plan:
- deactivate // to exist the current virtual environment
- sudo pip install virtualenv==20.0.23
- virtualenv -q --python=`which python` --always-copy ~/.virtualenv/test-env
- . ~/.virtualenv/test-env/bin/activate
- see that we're in the new 'test-env' virtual environment
- deactivate

NOTE: the reason why I didn't just run `make` is that the XCode part of dotfiles doesn't work for me. Also, I've had a bunch of other python issues with my local dev environment and I was worried that something unrelated to this fix might muck up my dev environment so I'm playing it safe.

Reviewers: john, davidbraley, benkraft, csilvers

Reviewed By: benkraft

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D64888