Ensure mobile repo submodules are up to date
Summary:
Today we ran into an issue where Android Studio wasn't seeing the style
config file that lives in the mobile repo. We finally tracked it down to
the fact that the mobile repo submodules were not initialized/up-to-date
(and the style config lives in a submodule).

This prevented Android Studio from seeing the file because it wasn't
even there (although curiously `ln -s` doesn't seem to complain if you
symlink to a non-existant file).

Issue: "none"

Test Plan:
1. Delete mobile repo
2. Delete the code style file: `rm -rf rm -f
   $HOME/Library/Preferences/AndroidStudio*/codestyles/KhanAcademyAndroid.xml`
3. Run `mac-android-setup.sh`

The repo should get cloned and when you launch Android Studio you should
be able to select the **KhanAcademyAndroid** style in: **Preferences** > **Editor** > **Code Style** > **Scheme**.

Reviewers: hannah

Reviewed By: hannah

Subscribers: csilvers, #mobile

Differential Revision: https://phabricator.khanacademy.org/D65471