Make sure we can actually install the new npm.
If you have node12 and try to run `npm install npm@6` when your npm is
5, it fails with some error involving 'cb'.  Luckily the default npm
on ubuntu (at least since 18.04) is npm 6.14, so just reinstalling the
default is enough to fix the problem.

It looks like nodejs can be installed in lots of different ways; this
works on my machine and will hopefully work on others' too!

Auditors: ragini

Test plan:
Ran this manually