Very initial packer template
This is just a placeholder for building a VM with linux-setup.sh.
The first step (this one) is to put the machinery around
programatically and consistently building a VM image.

Future steps may heavily modify some of the files here. And
we very much want to preserve those diffs.

Issue: https://khanacademy.atlassian.net/browse/INFRA-5638

Test plan:

Follow the instructions in packer_templates/README.md to build
the vagrant box. Then follow the instructions to launch the
box. Verify that it comes up.