Fix merge-base of system_report.sh
There were two errors in the git command, one of which spat out an error the shell, and the other made the output unreliable.
This fixes that, and improves it by providing the url of the diff rather than just the hash

Issue: "none"

Test plan:
Created a branch off of and old copy of webapp master, added a few commits, then ran the script. Saw no errors, and further observed that the provided info `Diverged from master at: https://github.com/Khan/webapp/commit/c7c898e594f 2021-06-30` was the LAST commit on master that was also on my branch (ie the common base).