INFRA-6889: Fix slow shell startup (#39)
Cache the homebrew prefix in ~/.brew_prefix so that shell does not take a long time to startup.

This problem came up because OS X x86 uses /usr/local and M1 uses /opt (and if a user wanted to they could specify their own location). Regardless, OS X defaults to having /usr/local/bin in the PATH, but it does not default to having /opt/(whatever) in the PATH. Thus, many things were breaking on M1s. The initial fix looked up this prefix on every single start of a shell and whatever brew is doing behind the scenes is slow. So now we cache the prefix location and shells startup faster.

Issue: https://khanacademy.atlassian.net/browse/INFRA-6889

Test plan:
I don't have a M1, so fingers crossed