macOS iOS Setup: install the xcodes utility (#42)
## Summary:

We've switched to recommending the `xcodes` utility to manage Xcode
installations. This tool allows us to install a specific version of Xcode
and not have it auto upgraded (which the App Store version does). 

It has the added benefit of not having any dependencies and no runtime 
requirements (ie. it doesn't depend on a scripting engine such as ruby or node)
and so is a single-executable install. This simplifies installation and
upgrades tremendously.

Issue: "none"

## Test plan:

Run `mac-ios-setup.sh`.
Once completed, you should be able to run `xcodes list` and see all of the 
available versions of Xcode.

Author: jeremywiebe

Reviewers: lillialexis, dbraley, jaredly, yogieric

Required Reviewers:

Approved By: lillialexis

Checks:

Pull Request URL: https://github.com/Khan/khan-dotfiles/pull/42