install nss and mkcert as part of linux-setup.sh (#46)
## Summary:
I recently made a tweak to install-mac-mkcert.py and assumed that no linux change was necessary because I couldn't find anything in the linux setup that was installing mkcert.  I assumed that this tool was part of the default ubuntu install.  How wrong I was.

This PR rectifies the situation by installing libnss3-tools which is a pre-req for mkcert and then builds mkcert for source before installing it.  The reason why I went the source route instead of downloading a pre-built binary from GitHub was to avoid potential architecture differences.  Also, the compile time is neglible.

I thought about reusing this function for the MacOS setup as well, but install_mkcert uses sudo for Linux, but on MacOS we'd want to run it without sudo, so ended up leaving the MacOS setup the way it is.

Issue: none

## Test plan:
- `sudo rm /usr/local/bin/mkcert` (from previous testing)
- copy the `install_mkcert` function, definition of `DEVTOOLS_DIR`, add a shebang, and a call to `install_mkcert` at the bottom
- `chmod 755 install_mkcert.sh`
- `./install_mkcert.sh`
- see it build and install mkcert
- `./install_mkcert.sh`
- see the `mkcert already installed` echoed
- run `mkcert -install` and see the following output (I've run it previous and don't know how to clear the trust store)
```
mkcert -install              
The local CA is already installed in the system trust store! 👍
The local CA is already installed in the Firefox trust store! 👍
```
- from webapp run `make -B genfiles/tlsproxy.pem` and see the following output:
```
mkcert \
		-cert-file genfiles/tlsproxy.pem \
		-key-file genfiles/tlsproxy-key.pem \
		"khanacademy.dev" "*.khanacademy.dev" \
		"khanacademy.test" "*.khanacademy.test" \
		"khanacademy.localhost" "*.khanacademy.localhost" \
		localhost 127.0.0.1 "::1"

Created a new certificate valid for the following names 📜
 - "khanacademy.dev"
 - "*.khanacademy.dev"
 - "khanacademy.test"
 - "*.khanacademy.test"
 - "khanacademy.localhost"
 - "*.khanacademy.localhost"
 - "localhost"
 - "127.0.0.1"
 - "::1"

Reminder: X.509 wildcards only go one level deep, so this won't match a.b.khanacademy.dev ℹ️

The certificate is at "genfiles/tlsproxy.pem" and the key at "genfiles/tlsproxy-key.pem" ✅

It will expire on 17 May 2024 🗓

# sync-end:dev-domains
```

Author: kevinbarabash

Reviewers: kevinbarabash, csilvers, yogieric

Required Reviewers:

Approved By: csilvers, yogieric

Checks:

Pull Request URL: https://github.com/Khan/khan-dotfiles/pull/46